changed:修改内存提前释放和边界值判断问题

+++ suuweibo/FontLabel/RTLabel.m
@@ -881,7 +881,7 @@
         [scanner scanUpToString:@">" intoString:&text];
         
         NSString *delimiter = [NSString stringWithFormat:@"%@>", text];
-        int position = [data rangeOfString:delimiter].location;
+        NSUInteger position = [data rangeOfString:delimiter].location;
         if (position!=NSNotFound)
         {
             if ([delimiter rangeOfString:@"<p"].location==0)
